Removed interface SiteDAO
